:art: Implement ``clang-tidy`` suggestions.

+++ include/fiction/technology/charge_distribution_surface.hpp
@@ -915,7 +915,8 @@ class charge_distribution_surface<Lyt, false> : public Lyt
         double defect_energy = 0;
         for (const auto& [cell, pot] : strg->defect_local_pot)
         {
-            defect_energy += pot * static_cast<double>(charge_state_to_sign(strg->cell_charge[cell_to_index(cell)]));
+            defect_energy += pot * static_cast<double>(charge_state_to_sign(
+                                       strg->cell_charge[static_cast<uint64_t>(cell_to_index(cell))]));
         }
 
         double defect_interaction = 0;
@@ -2030,16 +2031,17 @@ class charge_distribution_surface<Lyt, false> : public Lyt
                 if (counter_negative != dependent_cell_index_negative)
                 {
                     const auto sign = sign_to_charge_state(static_cast<int8_t>(remainder_int - 1));
-                    if (const auto new_chargesign = this->get_charge_state_by_index(
-                            static_cast<uint64_t>(cell_to_index(index_to_two_state_cell(counter_negative))));
+                    if (const auto new_chargesign = this->get_charge_state_by_index(static_cast<uint64_t>(
+                            cell_to_index(index_to_two_state_cell(static_cast<uint64_t>(counter_negative)))));
                         new_chargesign != sign)
                     {
-                        strg->cell_history.emplace_back(
-                            static_cast<uint64_t>(cell_to_index(index_to_two_state_cell(counter_negative))),
-                            charge_state_to_sign(new_chargesign));
+                        strg->cell_history.emplace_back(static_cast<uint64_t>(cell_to_index(index_to_two_state_cell(
+                                                            static_cast<uint64_t>(counter_negative)))),
+                                                        charge_state_to_sign(new_chargesign));
                         this->assign_charge_state_by_cell_index(
-                            static_cast<uint64_t>(cell_to_index(index_to_two_state_cell(counter_negative))), sign,
-                            false);
+                            static_cast<uint64_t>(
+                                cell_to_index(index_to_two_state_cell(static_cast<uint64_t>(counter_negative)))),
+                            sign, false);
                     }
                     counter_negative -= 1;
                 }
@@ -2049,15 +2051,16 @@ class charge_distribution_surface<Lyt, false> : public Lyt
                 {
                     counter_negative -= 1;
                     const auto sign = sign_to_charge_state(static_cast<int8_t>(remainder_int - 1));
-                    if (const auto old_chargesign =
-                            this->get_charge_state_by_index(cell_to_index(index_to_two_state_cell(counter_negative)));
+                    if (const auto old_chargesign = this->get_charge_state_by_index(
+                            cell_to_index(index_to_two_state_cell(static_cast<uint64_t>(counter_negative))));
                         old_chargesign != sign)
                     {
-                        strg->cell_history.emplace_back(
-                            static_cast<uint64_t>(cell_to_index(index_to_two_state_cell(counter_negative))),
-                            charge_state_to_sign(old_chargesign));
+                        strg->cell_history.emplace_back(static_cast<uint64_t>(cell_to_index(index_to_two_state_cell(
+                                                            static_cast<uint64_t>(counter_negative)))),
+                                                        charge_state_to_sign(old_chargesign));
                         this->assign_charge_state_by_cell_index(
-                            cell_to_index(index_to_two_state_cell(counter_negative)), sign, false);
+                            cell_to_index(index_to_two_state_cell(static_cast<uint64_t>(counter_negative))), sign,
+                            false);
                     }
                     counter_negative -= 1;
                 }
